FIX: prepare_data2 separator, ADD: dss-opt, info-rna, mcts-rna, ADD: work division in desi, inverse, redprint, sfbinv

+++ run_rnainverse.py
@@ -2,14 +2,29 @@ import os
 import sys
 import subprocess
 
-for ver in ['rnainverse', 'rnainverse_extended']:
+if len(sys.argv) == 3:
+    range_s = int(sys.argv[1])
+    range_e = int(sys.argv[2])
+else:
+    range_s = 0
+    range_e = 20000
 
-    indir = f'/rna_design/inputs/{ver}'
-    outdir = f'/rna_design/outputs/{ver}'
+dirs = [('/rna_design/inputs/rnainversed', '/rna_design/outputs/rnainverse'), 
+        ('/rna_design/inputs/rnainverse_extended', '/rna_design/outputs/rnainverse_extended'),
+        ('/rna_design/inputs2/rnainverse', '/rna_design/outputs2/rnainverse')]
 
-    for fn in os.listdir(indir):
+for indir, outdir in dirs:
+
+    to_do = os.listdir(indir)
+    to_do.sort()
+    to_do = to_do[range_s:range_e]
+    for fn in to_do:
         print(fn)
         tmp = fn.split('.')[0]
+        
+        if os.path.isfile(f'{outdir}/{tmp}.inv.out'):
+            continue
+        
         outfile = open(f'{outdir}/{tmp}.inv.out', 'w')
         errfile = open(f'{outdir}/{tmp}.inv.err', 'w')
